+++ src/JSystem/JKernel/JKRDvdAramRipper.cpp
@@ -103,7 +103,7 @@ JKRADCommand* JKRDvdAramRipper::callCommand_Async(JKRADCommand* command) {
         fileSize = ALIGN_NEXT(fileSize, 0x20);
         if (command->mExpandSwitch == 1) {
             u8 buffer[0x40];
-            u8* bufPtr = (u8*)ALIGN_NEXT((u32)&buffer, 0x20);
+            u8* bufPtr = (u8*)ALIGN_NEXT((uintptr_t)&buffer, 0x20);
             while (true) {
                 s32 result = DVDReadPrio(dvdFile->getFileInfo(), bufPtr, 0x20, 0, 2);
                 if (result >= 0) {
@@ -472,7 +472,7 @@ static u32 dmaBufferFlush(u32 param_1) {
         return 0;
     }
     u32 size = ALIGN_NEXT(dmaCurrent - dmaBuf, 0x20);
-    JKRAramPcs(0, (u32)dmaBuf, param_1, size, NULL);
+    JKRAramPcs(0, (uintptr_t)dmaBuf, param_1, size, NULL);
     dmaCurrent = dmaBuf;
     return size;
 }
